field.smooth adds smoothing length to a field.
asymetric kernel is used. (and in the different direction from
that used in raster). the result is good enough.

Comparin gto gadget's sml what we get here is systematcially
bigger.

There is a reference to the method that nishi showed me. He used
a similiar method in his sph code. The numerical solver used
here is a stupid one though.