generate the C code for reading the snapshot from a Reader object.
Note that the generated code are all static members.
Tthe caller shall wrap these functions with a public interface,
or feed them into a vtable, if multple formats are intended.

most useful functions are _read, and _write, and _pstart that returns
the starting index of a particle type within the block.

these functions return a string to the error argument if error occurs.
the string shall be freed by the caller.